Current status list works as intended, more to come
Reworked EffectStruct into a class instead so it could be serialized. Added function to remove statuses when foreach loop reducing duration finishes. Checks for stun and berserk are in Character.BeginTurn() and UIManager respectively. Application of heat statuses are not yet implemented, and cannot be removed naturally, will be integrated with heat