Polishing changes to desert and catacombs.
Attached PS_AmbientDust to Catacombs camera in correct position. Confirmed layering and setup of sprite for lighting purposes. (Sprite backside to cast shadows from behind, one sprite set to default layer for casting shadows from the front, one sprite in front of that set to a player layer with the light also set to player layer and culling mask set appropriately so the light is not cast on the floor.) Set up optional "harsh lighting" catacombs scene as a potential lighting option rather than the normal catacombs scene.